auth_server/requirements.txt
django
django-extensions
django-oauth-toolkit
django-sslserver
requests
PyJWT
gunicorn
cryptography